+++ packages/app-headless-cms-scheduler/src/components/Browser/CellLive.tsx
@@ -1,57 +1,15 @@
 import React, { useEffect, useRef } from "react";
 import { observer } from "mobx-react-lite";
-import { ReactComponent as ScheduledIcon } from "@webiny/icons/access_time.svg";
-import { Tag, Tooltip } from "@webiny/admin-ui";
 import { useContainer } from "@webiny/app";
 import { useModel } from "@webiny/app-headless-cms/exports/admin/cms.js";
 import { ContentEntryListConfig } from "@webiny/app-headless-cms/admin/config/contentEntries/index.js";
 import { ListScheduledActionsGateway } from "@webiny/app-scheduler/features/listScheduledActions/abstractions.js";
-import { ScheduleActionType } from "@webiny/app-scheduler/types.js";
-import type { SchedulerEntry } from "@webiny/app-scheduler/types.js";
-import { createNamespace, formatScheduledDate } from "~/utils/index.js";
+import { createNamespace } from "~/utils/index.js";
 import { scheduledActionsStore } from "./ScheduledActionsStore.js";
 import { schedulerMutationSignal } from "../schedulerMutationSignal.js";
-
-const scheduledTooltip = (scheduled: SchedulerEntry): string => {
-    const goLiveOn = scheduled.publishOn || scheduled.unpublishOn;
-    const actionLabel =
-        scheduled.actionType === ScheduleActionType.unpublish ? "unpublish" : "publish";
-    return goLiveOn
-        ? `Scheduled to ${actionLabel} on ${formatScheduledDate(goLiveOn)}`
-        : `Scheduled to ${actionLabel}`;
-};
-
-const ScheduledTag = ({ scheduled }: { scheduled: SchedulerEntry }) => (
-    <Tooltip
-        content={scheduledTooltip(scheduled)}
-        trigger={<Tag variant={"warning"} icon={<ScheduledIcon />} content={"Scheduled"} />}
-    />
-);
-
-const LiveTag = ({ version }: { version: number }) => (
-    <Tag swatchColor={"#5AC84C"} variant={"success-light"} content={`Live (v${version})`} />
-);
-
-/**
- * A scheduled action is superseded by a direct publish/unpublish. The API cancels it, but that
- * happens asynchronously, so reconcile against the entry's current state to clear the badge
- * immediately instead of waiting for the backend cancellation to propagate:
- *
- * - Scheduled PUBLISH is moot once the latest revision is already live (`live.version === version`).
- *   It is NOT moot when an older revision is live and a newer one is queued (live.version < version)
- *   — that is the legitimate "live and scheduled" case.
- * - Scheduled UNPUBLISH is moot once nothing is live.
- */
-const isScheduleMoot = (
-    scheduled: SchedulerEntry,
-    liveVersion: number | undefined,
-    currentVersion: number | undefined
-): boolean => {
-    if (scheduled.actionType === ScheduleActionType.unpublish) {
-        return liveVersion === undefined;
-    }
-    return liveVersion !== undefined && liveVersion === currentVersion;
-};
+import { LiveTag } from "./LiveTag.js";
+import { ScheduledTag } from "./ScheduledTag.js";
+import { isScheduleMoot } from "./isScheduleMoot.js";
 
 /**
  * Overrides the CMS "Live" column cell to surface scheduled publish/unpublish actions.
