+++ packages/app-headless-cms-scheduler/src/components/Browser/LiveTag.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Tag } from "@webiny/admin-ui";
+
+export const LiveTag = ({ version }: { version: number }) => (
+    <Tag swatchColor={"#5AC84C"} variant={"success-light"} content={`Live (v${version})`} />
+);
